--bugfix-- setting auto margin for flex col breaks sticky in ffox (works in chrome)
With a flex-flow, auto margins push the content to the extreme avail position.

The last commit works fine in chrome (desktop + mobile). But the auto margin for the sticky sidebar breaks the sticky feature in firefox (both on desktop and also on mobile).

Patched to keep same top and right margins without specifying the other two. Works on desktop for ffox + chrome. About to test on mobile.