[v8] +Tags liquid template + jekyll includes, fancy tag styles
Added a page listing each tag
& the posts that reference each one.

Tags are auto-populated by reading front-matter in jekyll posts under collection: Philosophy (folder /_Philosophy/..)

Two includes added, one as a generation method to create card previews of posts by reusing the title code of blog post page & reformatting it to fit in a nice flex row.

Tags are clickable and animate upon click.

Cards of posts nudge on hover and react by lowering opacity when clicked.

2nd include method places tags under the post card. Also used for placing under the post page's title.

A liquid {{ include.param }} is used as a switch to add the css style depending on adding to a preview card or adding to a page title.

So happy & tired. Awesome progress. Almost done.